OCR_TIKA/requirements.txt
requests==2.27.1
DateTime==4.3
kafka-python==2.0.2
termcolor==1.1.0
